requirements.txt
beautifulsoup4==4.11.1
black==22.12.0
click==8.1.3
colorama==0.4.6
lxml==4.9.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2==2.9.5
python-dotenv==0.21.1
soupsieve==2.3.2.post1
tomli==2.0.1
